// src/pokedex.types.ts
export interface Pokemon {
    abilities: Ability[];
    base_experience: number;
    cries: Cries;
    forms: Form[];
    game_indices: Index[];
    height: number;
    held_items: HeldItem[];
    id: number;
    is_default: boolean;
    location_area_encounters: string;
    moves: Mfe[];
    name: string;
    order: number;
    past_abilities: PastAbility[];
    past_stats: PastStat[];
    past_types: unknown[];
    species: Species;
    sprites: Sprites;
    stats: Stat3[];
    types: Type[];
    weight: number;
}

export interface Ability {
    ability: Ability2;
    is_hidden: boolean;
    slot: number;
}

export interface Ability2 {
    name: string;
    url: string;
}

export interface Cries {
    latest: string;
    legacy: string;
}

export interface Form {
    name: string;
    url: string;
}

export interface Index {
    game_index: number;
    version: Version;
}

export interface Version {
    name: string;
    url: string;
}

export interface HeldItem {
    item: Item;
    version_details: VersionDetail[];
}

export interface Item {
    name: string;
    url: string;
}

export interface VersionDetail {
    rarity: number;
    version: Version2;
}

export interface Version2 {
    name: string;
    url: string;
}

export interface Mfe {
    move: Move;
    version_group_details: VersionGroupDetail[];
}

export interface Move {
    name: string;
    url: string;
}

export interface VersionGroupDetail {
    level_learned_at: number;
    move_learn_method: MoveLearnMethod;
    order?: number;
    version_group: VersionGroup;
}

export interface MoveLearnMethod {
    name: string;
    url: string;
}

export interface VersionGroup {
    name: string;
    url: string;
}

export interface PastAbility {
    abilities: Ability3[];
    generation: Generation;
}

export interface Ability3 {
    ability: any;
    is_hidden: boolean;
    slot: number;
}

export interface Generation {
    name: string;
    url: string;
}

export interface PastStat {
    generation: Generation2;
    stats: Stat[];
}

export interface Generation2 {
    name: string;
    url: string;
}

export interface Stat {
    base_stat: number;
    effort: number;
    stat: Stat2;
}

export interface Stat2 {
    name: string;
    url: string;
}

export interface Species {
    name: string;
    url: string;
}

export interface Sprites {
    back_default: string;
    back_female: string;
    back_shiny: string;
    back_shiny_female: string;
    front_default: string;
    front_female: string;
    front_shiny: string;
    front_shiny_female: string;
    other: Other;
    versions: Versions;
}

export interface Other {
    dream_world: DreamWorld;
    home: Home;
    "official-artwork": OfficialArtwork;
    showdown: Showdown;
}
// ...
export interface Versions {
    "generation-i": GenerationI;
    "generation-ii": GenerationIi;
    "generation-iii": GenerationIii;
    "generation-iv": GenerationIv;
    "generation-ix": GenerationIx;
    "generation-v": GenerationV;
    "generation-vi": GenerationVi;
    "generation-vii": GenerationVii;
    "generation-viii": GenerationViii;
}
// ...
export interface Stat3 {
    base_stat: number;
    effort: number;
    stat: Stat4;
}

export interface Stat4 {
    name: string;
    url: string;
}

export interface Type {
    slot: number;
    type: Type2;
}

export interface Type2 {
    name: string;
    url: string;
}